% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNonZero.R
\name{countNonZero}
\alias{countNonZero}
\title{Non-zero element counter
This function returns the number of non-zero elements in a vector
(including values that can be coerced to equal 0, such as FALSE and "0").
It is also capable of handling non-vector inputs, such as lists and dataframes.
Values of NA will throw an error as they cannot be assigned a value of true/false.}
\usage{
countNonZero(vec)
}
\arguments{
\item{vec}{A vector or dataframe, featuring possible zeros.}
}
\value{
numNonZero An integer count of non zero values
}
\description{
Non-zero element counter
This function returns the number of non-zero elements in a vector
(including values that can be coerced to equal 0, such as FALSE and "0").
It is also capable of handling non-vector inputs, such as lists and dataframes.
Values of NA will throw an error as they cannot be assigned a value of true/false.
}
\examples{
df <- data.frame("ID"=as.character(seq(1,5)), "First_Aid_Training"=c(0,1,1,0,1))
# You can use countNonZero on most vectors, including numeric, character, and boolean
countNonZero(seq(0,5))
countNonZero(c("1","0","0","1"))
countNonZero(c(TRUE, FALSE, TRUE, TRUE))

# Alternatively, you can supply a dataframe as an argument as well.
# This will calculate the number of non-zero entries in each column and return a vector.
countNonZero(df)

}
