% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNonZero.R
\name{countNonZero}
\alias{countNonZero}
\title{Non-zero element counter
This function returns the number of non-zero elements in a vector
(including values that can be coerced to equal 0, such as FALSE and "0").
It is also capable of handling non-vector inputs, such as lists and dataframes.
Values of NA will throw an error as they cannot be assigned a value of true/false.}
\usage{
countNonZero(vec)
}
\arguments{
\item{vec}{A vector of inputs.
The name suggests that this function is designed to operate on generic vectors,
and not explicitly designed for higher level structures.}
}
\value{
numNonZero An integer count of non zero values
}
\description{
Non-zero element counter
This function returns the number of non-zero elements in a vector
(including values that can be coerced to equal 0, such as FALSE and "0").
It is also capable of handling non-vector inputs, such as lists and dataframes.
Values of NA will throw an error as they cannot be assigned a value of true/false.
}
\examples{
countNonZero(seq(0,5))
}
